+++ extension/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to the "al-translation-cleaner" extension will be documented
 
 <!-- Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file. -->
 
+## [Unreleased]
+
+### Fixed
+
+* translations were sometimes directly taken from the .g.xlf note, not the translation files
+
 ## [1.0.2] - 2026-03-30
 
 ### Added